added notice auth

+++ routes/notice.js
@@ -2,7 +2,9 @@ const express=require('express');
 const router = express.Router({ mergeParams: true });
 const catchAsync = require("../utils/catchAsync");
 const Notice = require("../models/notices");
-const { isLoggedIn } = require('../middleware');
+const { isLoggedIn, isComAd } = require('../middleware');
+const ExpressError = require("../utils/ExpressError");
+
 
 //Displays all the notices
 router.get('/',isLoggedIn, catchAsync(async (req,res,next)=>{
@@ -11,17 +13,17 @@ router.get('/',isLoggedIn, catchAsync(async (req,res,next)=>{
 }))
 
 //Displays the form to add a new notice
-router.get('/new',isLoggedIn, (req,res) => {
+router.get('/new',isLoggedIn, isComAd, (req,res) => {
     res.render('notices/new');
 })
 
 
 //Save the new notice in the database
-router.post('/', isLoggedIn,catchAsync(async (req,res,next) => {
-
-    const notice = new Notice(req.body.notices);
-    await notice.save();
-    res.redirect('/notices');
+router.post('/', isLoggedIn,isComAd, catchAsync(async (req,res,next) => {
+        const notice = new Notice(req.body.notices);
+        notice.author = req.user._id;
+        await notice.save();
+        res.redirect('/notices');
 }))
 
 //Show a particular Notice
@@ -32,23 +34,39 @@ router.get('/:id',isLoggedIn, catchAsync(async (req,res,next)=>{
 }))
 
 //to show the edit form
-router.get('/:id/edit', isLoggedIn, catchAsync(async (req,res,next) => {
+router.get('/:id/edit', isLoggedIn, isComAd, catchAsync(async (req,res,next) => {
     const { id } = req.params;
     const notice = await Notice.findById(id);
+    if((req.user._id.equals(notice.author)) || (req.user.user_role === 'Admin'))
     res.render('notices/edit', { notice });
+    else
+    throw new ExpressError("You are not not allowed access this page",401);
 }))
 
 //to save the edited details
-router.put('/:id',isLoggedIn, catchAsync(async (req,res,next)=>{
+router.put('/:id',isLoggedIn, isComAd, catchAsync(async (req,res,next)=>{
     const { id } = req.params;
-    const notice = await Notice.findByIdAndUpdate(id, {...req.body.notices});
-    res.redirect(`/notices/${notice._id}`);
+    const notice = await Notice.findById(id);
+    if((req.user._id.equals(notice.author)) || (req.user.user_role === 'Admin'))
+    {
+        const noti = await Notice.findByIdAndUpdate(id, {...req.body.notices});
+        res.redirect(`/notices/${notice._id}`);
+    }
+    else
+    throw new ExpressError("You are not not allowed access this page",401);
 }))
 
 //to delete the notice
 router.delete('/:id', catchAsync(async (req,res,next) => {
     const { id } = req.params;
-    await Notice.findByIdAndDelete(id);
-    res.redirect('/notices');
+    const notice = await Notice.findById(id);
+    if((req.user._id.equals(notice.author)) || (req.user.user_role === 'Admin'))
+    {
+        await Notice.findByIdAndDelete(id);
+        res.redirect('/notices');
+    }
+    else
+    throw new ExpressError("You are not not allowed access this page",401);
+
 }))
 module.exports = router;
